DriveIQ/DriveIQ-backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.112.2
fiona==1.10.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
fpdf==1.7.2
geopandas==0.14.4
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
haversine==2.8.1
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.2
networkx==3.2.1
numpy==1.26.4
openpyxl==3.1.5
osmnx==1.9.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.1
reportlab==4.2.5
requests==2.32.3
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.6
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
Werkzeug==3.0.4
wsproto==1.2.0
zipp==3.20.1
